chore: extract internal node setup
