+++ testing-tools/src/controllers/strom/mod.rs
@@ -2,6 +2,8 @@ mod node;
 pub use node::*;
 mod op_node;
 pub use op_node::*;
+mod op_internals;
+pub use op_internals::*;
 mod internals;
 pub use internals::*;
 mod harness;
